CategoryApi, CategoryService and CategoryApiImpl, CategoryServiceImpl

+++ src/main/java/com/graysan/business/services/ICategoryServices.java
@@ -28,6 +28,6 @@ public interface ICategoryServices<D,E> {
     // All Delete
     public String categoryServiceAllDelete();
     // Speed Data
-    public List<CategoryDto> categoryServiceSpeedData(int key);
+    public List<CategoryDto> categoryServiceSpeedData(Long key);
 
 }
