+++ src/main/java/com/graysan/business/services/impl/CategoryServiceImpl.java
@@ -124,7 +124,7 @@ public class CategoryServiceImpl implements ICategoryServices<CategoryDto, Categ
 
     // Speed Data
     @Override
-    public List<CategoryDto> categoryServiceSpeedData(int key) {
+    public List<CategoryDto> categoryServiceSpeedData(Long key) {
         CategoryDto categoryDto = null;
         List<CategoryDto> categoryDtoList=new ArrayList<>();
         int count = 0;
