+++ src/main/java/com/graysan/controller/api/ICategoryApi.java
@@ -31,6 +31,6 @@ public interface ICategoryApi<D> {
     public ResponseEntity<String> categoryApiAllDelete();
 
     // SPEED DATA
-    public  ResponseEntity<List<CategoryDto>> categoryApiSpeedData(int key);
+    public  ResponseEntity<List<CategoryDto>> categoryApiSpeedData(Long key);
 
 }
